% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_deg.R
\name{run_DEG}
\alias{run_DEG}
\title{Run differential expression and TE analysis}
\usage{
run_DEG(seurat_obj, cohort_name, cluster_name, metadata_df, outdir = ".")
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{cohort_name}{Character}

\item{cluster_name}{Character}

\item{metadata_df}{Data frame with gene annotations}

\item{outdir}{Output directory (default = current working dir)}
}
\description{
Run differential expression and TE analysis
}
